Allow float values for team change ramp-up periods
This change enables more granular control over team member ramp-up periods
by accepting float values instead of only integers.

Changes:
- Modified VelocityAdjustmentParser to parse ramp as float
- Updated TeamChange dataclass to accept float for ramp_up_sprints
- Added test case for float ramp values
- Updated README documentation to clarify float support

This allows scenarios like 2.5 sprint ramp-up periods for more accurate
velocity modeling when onboarding new team members.

+++ src/domain/velocity_adjustments.py
@@ -67,7 +67,7 @@ class TeamChange:
 
     sprint: int
     change: float  # +2 for additions, -2 for departures, +0.5 for part-time
-    ramp_up_sprints: int = 3
+    ramp_up_sprints: float = 3.0
     productivity_curve: ProductivityCurve = ProductivityCurve.LINEAR
 
     def get_productivity_factor(self, sprints_since_change: int) -> float:
